requirements.txt
aiogram==3.4.1
google-generativeai==0.5.1
python-dotenv==1.0.1
